specs: Refactor MockWebRequest and update SignInRequestTest
